requirements.txt
asgiref==3.7.2
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-jsonapi==6.1.0
greenlet==3.0.1
inflection==0.5.1
mysqlclient==2.2.0
psycopg2==2.9.9
PyMySQL==1.1.0
pytz==2023.3.post1
SQLAlchemy==2.0.23
sqlparse==0.4.4
typing_extensions==4.8.0
